+++ dolfin/refinement/ParallelRefinement.cpp
@@ -194,7 +194,7 @@ void ParallelRefinement::create_new_vertices()
   const std::size_t num_new_vertices = n;
   const std::size_t global_offset
     = MPI::global_offset(_mesh.mpi_comm(), num_new_vertices, true)
-    + _mesh.size_global(0);
+    + _mesh.num_entities_global(0);
 
   // If they are shared, then the new global vertex index needs to be
   // sent off-process.  Add offset to map, and collect up any shared
@@ -254,7 +254,15 @@ void ParallelRefinement::build_local(Mesh& new_mesh) const
   dolfin_assert(new_cell_topology.size()%num_cell_vertices == 0);
   const std::size_t num_cells = new_cell_topology.size()/num_cell_vertices;
 
-  ed.open(new_mesh, tdim, gdim);
+  CellType::Type cell_type;
+  if (tdim == 3)
+    cell_type = CellType::tetrahedron;
+  else if (tdim == 2)
+    cell_type = CellType::triangle;
+  else
+    cell_type = CellType::interval;
+
+  ed.open(new_mesh, cell_type, tdim, gdim);
   ed.init_vertices(num_vertices);
   std::size_t i = 0;
   for (auto p = new_vertex_coordinates.begin();