+++ python/src/adaptivity.cpp
@@ -31,6 +31,7 @@
 #include <dolfin/fem/Form.h>
 #include <dolfin/fem/LinearVariationalProblem.h>
 #include <dolfin/fem/NonlinearVariationalProblem.h>
+#include <dolfin/function/Function.h>
 #include <dolfin/la/GenericVector.h>
 #include <dolfin/mesh/Mesh.h>
 
@@ -45,7 +46,7 @@ namespace dolfin_wrappers
   {
 #ifdef HAS_HDF5
     // dolfin::TimesSeries
-    py::class_<dolfin::TimeSeries, std::shared_ptr<dolfin::TimeSeries>>(m, "TimeSeries")
+    py::class_<dolfin::TimeSeries, std::shared_ptr<dolfin::TimeSeries>, dolfin::Variable>(m, "TimeSeries")
       .def(py::init<std::string>())
       .def(py::init([](const MPICommWrapper comm, const std::string &arg)
                     { return std::unique_ptr<dolfin::TimeSeries>(new dolfin::TimeSeries(comm.get(), arg)); }))