+++ test/unit/python/fem/test_discrete_operators.py
@@ -34,9 +34,9 @@ def test_gradient():
         W = FunctionSpace(mesh, "Nedelec 1st kind H(curl)", 1)
 
         G = DiscreteOperators.build_gradient(W, V)
-        num_edges = mesh.size_global(1)
+        num_edges = mesh.num_entities_global(1)
         assert G.size(0) == num_edges
-        assert G.size(1) == mesh.size_global(0)
+        assert G.size(1) == mesh.num_entities_global(0)
         assert round(G.norm("frobenius") - sqrt(2.0*num_edges), 8) == 0.0
 
     meshes = [UnitSquareMesh(11, 6), UnitCubeMesh(4, 3, 7)]